# internal/services/media_service/media/features/restoring_file/v1/dtos/get_media_response_dto.go
package dtos

import (
	"github.com/meysamhadeli/shop-golang-microservices/internal/pkg/utils"
	"github.com/meysamhadeli/shop-golang-microservices/internal/services/product_service/media/dtos"
)

type GetMediasResponseDto struct {
	Medias *utils.ListResult[*dtos.MediaDto]
}