# internal/services/media_service/media/features/uploading_file/v1/dtos/create_media_request_dto.go
package dtos

type CreateMediaRequestDto struct {
	Name        string  `json:"name"`
	Description string  `json:"description"`
	Price       float64 `json:"price"`
	Count       int32   `json:"count"`
	InventoryId int64   `json:"inventoryId"`
}